:sparkles: Split upload_avatar into two methods.
We need to be able to create avatar objects with data as a data URI as
well as from an uploaded file. This adds a new method to perform the
operation as an upload as a data uri, and renames the previous method to a
distinct path that allows for upload as a raw file.