:zap: Remove data uris and images from avatar listing.
Our queries to list avatars return all avatars in the system. As such, if
we are sending large amounts of data (i.e. in the form of images), this
could end up being a VERY large amount of data sent across the wire. This
commit hides that data and doesn't send it as part of the initial listing
response. Instead, a user would have to make a subsequent query to get a
user's avatar.

Fixes #7.